Fix race condition in stack allocator
When freeing memory stack allocator sets all bytes (including the separator) to UNALLOC_BYTE. At the same time watchdog may check whether the separator is equal to 0x00FACADE and fail. The solution is to move the memset invocation after the loop, which decreases allocations_num. At this point allocations are already freed and the watchdog will not check them, so we may safely reset the memory.